+++ src/scene/light.rs
@@ -1,12 +1,12 @@
 #![forbid(unsafe_code)]
 
 pub struct Light {
-    pub position: Vec<f32>,
+    pub position: Vec<f64>,
     pub brightness: u8,
 }
 
 impl Light {
-    pub fn new(position: Vec<f32>, brightness: u8) -> Self {
+    pub fn new(position: Vec<f64>, brightness: u8) -> Self {
         Light {
             position,
             brightness,
